dev-console/requirements-dev-console.txt
# Development Minecraft Console - Dependencies
# Built following Elon Musk principles: simple, fast, effective

# UI Framework - Modern, gradient support
customtkinter==5.2.1

# Backend API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Real-time communication
websockets==12.0

# File watching for hot-reload
watchdog==3.0.0

# Version control integration
gitpython==3.1.40

# Authentication
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.1

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# HTTP requests
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pillow==10.1.0

# RCON for server control
mcrcon==0.7.0

# Mod metadata parsing
zipfile36==0.1.3

# AI Integration
python-dotenv==1.0.0

